4.0.2-线下还款

+++ 51fbapi-dal/src/main/java/com/ald/fanbei/api/dal/dao/AfBorrowLegalOrderDao.java
@@ -28,6 +28,6 @@ public interface AfBorrowLegalOrderDao extends BaseDao<AfBorrowLegalOrderDo, Lon
 	 * @param borrowId
 	 * @return
 	 */
-	Integer tuchByBorrowId(Long borrowId);
+	Long tuchByBorrowId(Long borrowId);
 
 }
